Restore logging and wrap runServer error
